pyproject.toml
[project]
name = "swiftly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.2",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
]
